resolved most of the issues raised by tom, incorperated control precision test, moving to collision test and grasping held test now
