+++ docs/envs/DynObstruction2D-o0.md
@@ -31,45 +31,41 @@ The entries of an array in this Box space correspond to the following object fea
 | 4 | target_surface | vy |
 | 5 | target_surface | omega |
 | 6 | target_surface | static |
-| 7 | target_surface | kinematic |
-| 8 | target_surface | dynamic |
-| 9 | target_surface | color_r |
-| 10 | target_surface | color_g |
-| 11 | target_surface | color_b |
-| 12 | target_surface | z_order |
-| 13 | target_surface | width |
-| 14 | target_surface | height |
-| 15 | target_block | x |
-| 16 | target_block | y |
-| 17 | target_block | theta |
-| 18 | target_block | vx |
-| 19 | target_block | vy |
-| 20 | target_block | omega |
-| 21 | target_block | static |
-| 22 | target_block | kinematic |
-| 23 | target_block | dynamic |
-| 24 | target_block | color_r |
-| 25 | target_block | color_g |
-| 26 | target_block | color_b |
-| 27 | target_block | z_order |
-| 28 | target_block | width |
-| 29 | target_block | height |
-| 30 | target_block | mass |
-| 31 | target_block | moment |
-| 32 | robot | x |
-| 33 | robot | vx |
-| 34 | robot | y |
-| 35 | robot | vy |
-| 36 | robot | theta |
-| 37 | robot | omega |
-| 38 | robot | base_radius |
-| 39 | robot | arm_joint |
-| 40 | robot | arm_length |
-| 41 | robot | gripper_base_width |
-| 42 | robot | gripper_base_height |
-| 43 | robot | finger_gap |
-| 44 | robot | finger_height |
-| 45 | robot | finger_width |
+| 7 | target_surface | color_r |
+| 8 | target_surface | color_g |
+| 9 | target_surface | color_b |
+| 10 | target_surface | z_order |
+| 11 | target_surface | width |
+| 12 | target_surface | height |
+| 13 | target_block | x |
+| 14 | target_block | y |
+| 15 | target_block | theta |
+| 16 | target_block | vx |
+| 17 | target_block | vy |
+| 18 | target_block | omega |
+| 19 | target_block | static |
+| 20 | target_block | color_r |
+| 21 | target_block | color_g |
+| 22 | target_block | color_b |
+| 23 | target_block | z_order |
+| 24 | target_block | width |
+| 25 | target_block | height |
+| 26 | target_block | mass |
+| 27 | robot | x |
+| 28 | robot | vx |
+| 29 | robot | y |
+| 30 | robot | vy |
+| 31 | robot | theta |
+| 32 | robot | omega |
+| 33 | robot | static |
+| 34 | robot | base_radius |
+| 35 | robot | arm_joint |
+| 36 | robot | arm_length |
+| 37 | robot | gripper_base_width |
+| 38 | robot | gripper_base_height |
+| 39 | robot | finger_gap |
+| 40 | robot | finger_height |
+| 41 | robot | finger_width |
 
 
 ### Action Space
