+++ docs/envs/DynObstruction2D-o1.md
@@ -33,60 +33,55 @@ The entries of an array in this Box space correspond to the following object fea
 | 4 | target_surface | vy |
 | 5 | target_surface | omega |
 | 6 | target_surface | static |
-| 7 | target_surface | kinematic |
-| 8 | target_surface | dynamic |
-| 9 | target_surface | color_r |
-| 10 | target_surface | color_g |
-| 11 | target_surface | color_b |
-| 12 | target_surface | z_order |
-| 13 | target_surface | width |
-| 14 | target_surface | height |
-| 15 | target_block | x |
-| 16 | target_block | y |
-| 17 | target_block | theta |
-| 18 | target_block | vx |
-| 19 | target_block | vy |
-| 20 | target_block | omega |
-| 21 | target_block | static |
-| 22 | target_block | kinematic |
-| 23 | target_block | dynamic |
-| 24 | target_block | color_r |
-| 25 | target_block | color_g |
-| 26 | target_block | color_b |
-| 27 | target_block | z_order |
-| 28 | target_block | width |
-| 29 | target_block | height |
-| 30 | target_block | mass |
-| 31 | target_block | moment |
-| 32 | obstruction0 | x |
-| 33 | obstruction0 | y |
-| 34 | obstruction0 | theta |
-| 35 | obstruction0 | vx |
-| 36 | obstruction0 | vy |
-| 37 | obstruction0 | omega |
-| 38 | obstruction0 | static |
-| 39 | obstruction0 | kinematic |
-| 40 | obstruction0 | dynamic |
-| 41 | obstruction0 | color_r |
-| 42 | obstruction0 | color_g |
-| 43 | obstruction0 | color_b |
-| 44 | obstruction0 | z_order |
-| 45 | obstruction0 | width |
-| 46 | obstruction0 | height |
-| 47 | robot | x |
-| 48 | robot | vx |
-| 49 | robot | y |
-| 50 | robot | vy |
-| 51 | robot | theta |
-| 52 | robot | omega |
-| 53 | robot | base_radius |
-| 54 | robot | arm_joint |
-| 55 | robot | arm_length |
-| 56 | robot | gripper_base_width |
-| 57 | robot | gripper_base_height |
-| 58 | robot | finger_gap |
-| 59 | robot | finger_height |
-| 60 | robot | finger_width |
+| 7 | target_surface | color_r |
+| 8 | target_surface | color_g |
+| 9 | target_surface | color_b |
+| 10 | target_surface | z_order |
+| 11 | target_surface | width |
+| 12 | target_surface | height |
+| 13 | target_block | x |
+| 14 | target_block | y |
+| 15 | target_block | theta |
+| 16 | target_block | vx |
+| 17 | target_block | vy |
+| 18 | target_block | omega |
+| 19 | target_block | static |
+| 20 | target_block | color_r |
+| 21 | target_block | color_g |
+| 22 | target_block | color_b |
+| 23 | target_block | z_order |
+| 24 | target_block | width |
+| 25 | target_block | height |
+| 26 | target_block | mass |
+| 27 | obstruction0 | x |
+| 28 | obstruction0 | y |
+| 29 | obstruction0 | theta |
+| 30 | obstruction0 | vx |
+| 31 | obstruction0 | vy |
+| 32 | obstruction0 | omega |
+| 33 | obstruction0 | static |
+| 34 | obstruction0 | color_r |
+| 35 | obstruction0 | color_g |
+| 36 | obstruction0 | color_b |
+| 37 | obstruction0 | z_order |
+| 38 | obstruction0 | width |
+| 39 | obstruction0 | height |
+| 40 | obstruction0 | mass |
+| 41 | robot | x |
+| 42 | robot | vx |
+| 43 | robot | y |
+| 44 | robot | vy |
+| 45 | robot | theta |
+| 46 | robot | omega |
+| 47 | robot | static |
+| 48 | robot | base_radius |
+| 49 | robot | arm_joint |
+| 50 | robot | arm_length |
+| 51 | robot | gripper_base_width |
+| 52 | robot | gripper_base_height |
+| 53 | robot | finger_gap |
+| 54 | robot | finger_height |
+| 55 | robot | finger_width |
 
 
 ### Action Space
