+++ docs/envs/DynObstruction2D-o2.md
@@ -33,75 +33,69 @@ The entries of an array in this Box space correspond to the following object fea
 | 4 | target_surface | vy |
 | 5 | target_surface | omega |
 | 6 | target_surface | static |
-| 7 | target_surface | kinematic |
-| 8 | target_surface | dynamic |
-| 9 | target_surface | color_r |
-| 10 | target_surface | color_g |
-| 11 | target_surface | color_b |
-| 12 | target_surface | z_order |
-| 13 | target_surface | width |
-| 14 | target_surface | height |
-| 15 | target_block | x |
-| 16 | target_block | y |
-| 17 | target_block | theta |
-| 18 | target_block | vx |
-| 19 | target_block | vy |
-| 20 | target_block | omega |
-| 21 | target_block | static |
-| 22 | target_block | kinematic |
-| 23 | target_block | dynamic |
-| 24 | target_block | color_r |
-| 25 | target_block | color_g |
-| 26 | target_block | color_b |
-| 27 | target_block | z_order |
-| 28 | target_block | width |
-| 29 | target_block | height |
-| 30 | target_block | mass |
-| 31 | target_block | moment |
-| 32 | obstruction0 | x |
-| 33 | obstruction0 | y |
-| 34 | obstruction0 | theta |
-| 35 | obstruction0 | vx |
-| 36 | obstruction0 | vy |
-| 37 | obstruction0 | omega |
-| 38 | obstruction0 | static |
-| 39 | obstruction0 | kinematic |
-| 40 | obstruction0 | dynamic |
-| 41 | obstruction0 | color_r |
-| 42 | obstruction0 | color_g |
-| 43 | obstruction0 | color_b |
-| 44 | obstruction0 | z_order |
-| 45 | obstruction0 | width |
-| 46 | obstruction0 | height |
-| 47 | obstruction1 | x |
-| 48 | obstruction1 | y |
-| 49 | obstruction1 | theta |
-| 50 | obstruction1 | vx |
-| 51 | obstruction1 | vy |
-| 52 | obstruction1 | omega |
-| 53 | obstruction1 | static |
-| 54 | obstruction1 | kinematic |
-| 55 | obstruction1 | dynamic |
-| 56 | obstruction1 | color_r |
-| 57 | obstruction1 | color_g |
-| 58 | obstruction1 | color_b |
-| 59 | obstruction1 | z_order |
-| 60 | obstruction1 | width |
-| 61 | obstruction1 | height |
-| 62 | robot | x |
-| 63 | robot | vx |
-| 64 | robot | y |
-| 65 | robot | vy |
-| 66 | robot | theta |
-| 67 | robot | omega |
-| 68 | robot | base_radius |
-| 69 | robot | arm_joint |
-| 70 | robot | arm_length |
-| 71 | robot | gripper_base_width |
-| 72 | robot | gripper_base_height |
-| 73 | robot | finger_gap |
-| 74 | robot | finger_height |
-| 75 | robot | finger_width |
+| 7 | target_surface | color_r |
+| 8 | target_surface | color_g |
+| 9 | target_surface | color_b |
+| 10 | target_surface | z_order |
+| 11 | target_surface | width |
+| 12 | target_surface | height |
+| 13 | target_block | x |
+| 14 | target_block | y |
+| 15 | target_block | theta |
+| 16 | target_block | vx |
+| 17 | target_block | vy |
+| 18 | target_block | omega |
+| 19 | target_block | static |
+| 20 | target_block | color_r |
+| 21 | target_block | color_g |
+| 22 | target_block | color_b |
+| 23 | target_block | z_order |
+| 24 | target_block | width |
+| 25 | target_block | height |
+| 26 | target_block | mass |
+| 27 | obstruction0 | x |
+| 28 | obstruction0 | y |
+| 29 | obstruction0 | theta |
+| 30 | obstruction0 | vx |
+| 31 | obstruction0 | vy |
+| 32 | obstruction0 | omega |
+| 33 | obstruction0 | static |
+| 34 | obstruction0 | color_r |
+| 35 | obstruction0 | color_g |
+| 36 | obstruction0 | color_b |
+| 37 | obstruction0 | z_order |
+| 38 | obstruction0 | width |
+| 39 | obstruction0 | height |
+| 40 | obstruction0 | mass |
+| 41 | obstruction1 | x |
+| 42 | obstruction1 | y |
+| 43 | obstruction1 | theta |
+| 44 | obstruction1 | vx |
+| 45 | obstruction1 | vy |
+| 46 | obstruction1 | omega |
+| 47 | obstruction1 | static |
+| 48 | obstruction1 | color_r |
+| 49 | obstruction1 | color_g |
+| 50 | obstruction1 | color_b |
+| 51 | obstruction1 | z_order |
+| 52 | obstruction1 | width |
+| 53 | obstruction1 | height |
+| 54 | obstruction1 | mass |
+| 55 | robot | x |
+| 56 | robot | vx |
+| 57 | robot | y |
+| 58 | robot | vy |
+| 59 | robot | theta |
+| 60 | robot | omega |
+| 61 | robot | static |
+| 62 | robot | base_radius |
+| 63 | robot | arm_joint |
+| 64 | robot | arm_length |
+| 65 | robot | gripper_base_width |
+| 66 | robot | gripper_base_height |
+| 67 | robot | finger_gap |
+| 68 | robot | finger_height |
+| 69 | robot | finger_width |
 
 
 ### Action Space
