+++ docs/envs/DynObstruction2D-o4.md
@@ -33,105 +33,97 @@ The entries of an array in this Box space correspond to the following object fea
 | 4 | target_surface | vy |
 | 5 | target_surface | omega |
 | 6 | target_surface | static |
-| 7 | target_surface | kinematic |
-| 8 | target_surface | dynamic |
-| 9 | target_surface | color_r |
-| 10 | target_surface | color_g |
-| 11 | target_surface | color_b |
-| 12 | target_surface | z_order |
-| 13 | target_surface | width |
-| 14 | target_surface | height |
-| 15 | target_block | x |
-| 16 | target_block | y |
-| 17 | target_block | theta |
-| 18 | target_block | vx |
-| 19 | target_block | vy |
-| 20 | target_block | omega |
-| 21 | target_block | static |
-| 22 | target_block | kinematic |
-| 23 | target_block | dynamic |
-| 24 | target_block | color_r |
-| 25 | target_block | color_g |
-| 26 | target_block | color_b |
-| 27 | target_block | z_order |
-| 28 | target_block | width |
-| 29 | target_block | height |
-| 30 | target_block | mass |
-| 31 | target_block | moment |
-| 32 | obstruction0 | x |
-| 33 | obstruction0 | y |
-| 34 | obstruction0 | theta |
-| 35 | obstruction0 | vx |
-| 36 | obstruction0 | vy |
-| 37 | obstruction0 | omega |
-| 38 | obstruction0 | static |
-| 39 | obstruction0 | kinematic |
-| 40 | obstruction0 | dynamic |
-| 41 | obstruction0 | color_r |
-| 42 | obstruction0 | color_g |
-| 43 | obstruction0 | color_b |
-| 44 | obstruction0 | z_order |
-| 45 | obstruction0 | width |
-| 46 | obstruction0 | height |
-| 47 | obstruction1 | x |
-| 48 | obstruction1 | y |
-| 49 | obstruction1 | theta |
-| 50 | obstruction1 | vx |
-| 51 | obstruction1 | vy |
-| 52 | obstruction1 | omega |
-| 53 | obstruction1 | static |
-| 54 | obstruction1 | kinematic |
-| 55 | obstruction1 | dynamic |
-| 56 | obstruction1 | color_r |
-| 57 | obstruction1 | color_g |
-| 58 | obstruction1 | color_b |
-| 59 | obstruction1 | z_order |
-| 60 | obstruction1 | width |
-| 61 | obstruction1 | height |
-| 62 | obstruction2 | x |
-| 63 | obstruction2 | y |
-| 64 | obstruction2 | theta |
-| 65 | obstruction2 | vx |
-| 66 | obstruction2 | vy |
-| 67 | obstruction2 | omega |
-| 68 | obstruction2 | static |
-| 69 | obstruction2 | kinematic |
-| 70 | obstruction2 | dynamic |
-| 71 | obstruction2 | color_r |
-| 72 | obstruction2 | color_g |
-| 73 | obstruction2 | color_b |
-| 74 | obstruction2 | z_order |
-| 75 | obstruction2 | width |
-| 76 | obstruction2 | height |
-| 77 | obstruction3 | x |
-| 78 | obstruction3 | y |
-| 79 | obstruction3 | theta |
-| 80 | obstruction3 | vx |
-| 81 | obstruction3 | vy |
-| 82 | obstruction3 | omega |
-| 83 | obstruction3 | static |
-| 84 | obstruction3 | kinematic |
-| 85 | obstruction3 | dynamic |
-| 86 | obstruction3 | color_r |
-| 87 | obstruction3 | color_g |
-| 88 | obstruction3 | color_b |
-| 89 | obstruction3 | z_order |
-| 90 | obstruction3 | width |
-| 91 | obstruction3 | height |
-| 92 | robot | x |
-| 93 | robot | vx |
-| 94 | robot | y |
-| 95 | robot | vy |
-| 96 | robot | theta |
-| 97 | robot | omega |
-| 98 | robot | base_radius |
-| 99 | robot | arm_joint |
-| 100 | robot | arm_length |
-| 101 | robot | gripper_base_width |
-| 102 | robot | gripper_base_height |
-| 103 | robot | finger_gap |
-| 104 | robot | finger_height |
-| 105 | robot | finger_width |
+| 7 | target_surface | color_r |
+| 8 | target_surface | color_g |
+| 9 | target_surface | color_b |
+| 10 | target_surface | z_order |
+| 11 | target_surface | width |
+| 12 | target_surface | height |
+| 13 | target_block | x |
+| 14 | target_block | y |
+| 15 | target_block | theta |
+| 16 | target_block | vx |
+| 17 | target_block | vy |
+| 18 | target_block | omega |
+| 19 | target_block | static |
+| 20 | target_block | color_r |
+| 21 | target_block | color_g |
+| 22 | target_block | color_b |
+| 23 | target_block | z_order |
+| 24 | target_block | width |
+| 25 | target_block | height |
+| 26 | target_block | mass |
+| 27 | obstruction0 | x |
+| 28 | obstruction0 | y |
+| 29 | obstruction0 | theta |
+| 30 | obstruction0 | vx |
+| 31 | obstruction0 | vy |
+| 32 | obstruction0 | omega |
+| 33 | obstruction0 | static |
+| 34 | obstruction0 | color_r |
+| 35 | obstruction0 | color_g |
+| 36 | obstruction0 | color_b |
+| 37 | obstruction0 | z_order |
+| 38 | obstruction0 | width |
+| 39 | obstruction0 | height |
+| 40 | obstruction0 | mass |
+| 41 | obstruction1 | x |
+| 42 | obstruction1 | y |
+| 43 | obstruction1 | theta |
+| 44 | obstruction1 | vx |
+| 45 | obstruction1 | vy |
+| 46 | obstruction1 | omega |
+| 47 | obstruction1 | static |
+| 48 | obstruction1 | color_r |
+| 49 | obstruction1 | color_g |
+| 50 | obstruction1 | color_b |
+| 51 | obstruction1 | z_order |
+| 52 | obstruction1 | width |
+| 53 | obstruction1 | height |
+| 54 | obstruction1 | mass |
+| 55 | obstruction2 | x |
+| 56 | obstruction2 | y |
+| 57 | obstruction2 | theta |
+| 58 | obstruction2 | vx |
+| 59 | obstruction2 | vy |
+| 60 | obstruction2 | omega |
+| 61 | obstruction2 | static |
+| 62 | obstruction2 | color_r |
+| 63 | obstruction2 | color_g |
+| 64 | obstruction2 | color_b |
+| 65 | obstruction2 | z_order |
+| 66 | obstruction2 | width |
+| 67 | obstruction2 | height |
+| 68 | obstruction2 | mass |
+| 69 | obstruction3 | x |
+| 70 | obstruction3 | y |
+| 71 | obstruction3 | theta |
+| 72 | obstruction3 | vx |
+| 73 | obstruction3 | vy |
+| 74 | obstruction3 | omega |
+| 75 | obstruction3 | static |
+| 76 | obstruction3 | color_r |
+| 77 | obstruction3 | color_g |
+| 78 | obstruction3 | color_b |
+| 79 | obstruction3 | z_order |
+| 80 | obstruction3 | width |
+| 81 | obstruction3 | height |
+| 82 | obstruction3 | mass |
+| 83 | robot | x |
+| 84 | robot | vx |
+| 85 | robot | y |
+| 86 | robot | vy |
+| 87 | robot | theta |
+| 88 | robot | omega |
+| 89 | robot | static |
+| 90 | robot | base_radius |
+| 91 | robot | arm_joint |
+| 92 | robot | arm_length |
+| 93 | robot | gripper_base_width |
+| 94 | robot | gripper_base_height |
+| 95 | robot | finger_gap |
+| 96 | robot | finger_height |
+| 97 | robot | finger_width |
 
 
 ### Action Space
